feat(api,web): add login and basic crud operations for book

+++ src/api/Commands/Book/DeleteBookCommand.cs
@@ -10,12 +10,14 @@ public record DeleteBookCommand(Guid Id) : IResultCommand<Unit>;
 public class DeleteBookCommandHandler : IResultCommandHandler<DeleteBookCommand, Unit>
 {
     private readonly IBookRepository _bookRepository;
+    private readonly IReservationRepository _reservationRepository;
 
     public DeleteBookCommandHandler(
-        IBookRepository bookRepository
-        )
+        IBookRepository bookRepository,
+        IReservationRepository reservationRepository)
     {
         _bookRepository = bookRepository;
+        _reservationRepository = reservationRepository;
     }
 
     public async Task<Result<Unit, Error>> Handle(DeleteBookCommand request, CancellationToken cancellationToken)
@@ -25,6 +27,14 @@ public class DeleteBookCommandHandler : IResultCommandHandler<DeleteBookCommand,
         {
             return BookErrors.NotFound;
         }
+
+        var hasAnyReservations = await _reservationRepository.HasAnyReservations(book.Id, cancellationToken);
+        if (hasAnyReservations)
+        {
+            return BookErrors.HasReservations;
+        }
+
+
         await _bookRepository.DeleteAsync(book.Id, cancellationToken);
         await _bookRepository.SaveChangesAsync(cancellationToken);
         return Unit.Value;
